+++ test/persistence/DummyMongoDbPersistence.py
@@ -12,7 +12,7 @@
 from pip_services3_commons.data import FilterParams
 
 from pip_services3_mongodb.persistence import IdentifiableMongoDbPersistence
-from ..IDummyPersistence import IDummyPersistence
+from test.fixtures.IDummyPersistence import IDummyPersistence
 
 
 class DummyMongoDbPersistence(IdentifiableMongoDbPersistence, IDummyPersistence):