+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> MongoDB components for Python Changelog
 
+## <a name="3.1.1"></a> 3.1.1 (2021-08-09)
+
+### Features
+* Added _auto_generate_id flag to IdentifiableMongoDbPersistence
+
 ## <a name="3.1.0"></a> 3.1.0 (2021-05-12)
 
 ### Bug Fixes